Add user verification to deleting blogs

+++ part4/bloglist/controllers/blogsRouter.js
@@ -38,7 +38,24 @@ blogsRouter.post("/", async (request, response, next) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
   const { id } = request.params;
+
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+
+    const blog = await Blog.findById(id);
+    if (!blog) {
+      return response.status(404).json({ error: "blog not found" });
+    }
+
+    if (blog.user.toString() !== decodedToken.id) {
+      return response
+        .status(403)
+        .json({ error: "not authorized to delete this blog" });
+    }
+
     await Blog.findByIdAndDelete(id);
     response.status(204).end();
   } catch (exception) {
